+++ printing/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.0.0-nullsafety.2
 
 - Add imageFromAssetBundle and networkImage
+- Add Page orientation on PdfPreview
 
 ## 5.0.0-nullsafety.1